refactor: Rename workspace to types
