+++ src/background/types.ts
